% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oxcal_formatter}
\alias{oxcal_formatter}
\title{read data in, three columns Name = lab code, Date = radiocarbon age, Uncertainty = error}
\usage{
oxcal_formatter(dates)
}
\arguments{
\item{dates}{raw data from Direct AMS lab}
}
\description{
read data in, three columns Name = lab code, Date = radiocarbon age, Uncertainty = error
}

